requirements.txt
pycocotools
torchvision==0.2.2.post3
torch==1.1.0
numpy==1.16.3
Pillow==6.2.1
